requirements.txt
langchain
langchain-openai
langchain-community
pinecone-client
tavily-python
tiktoken
pinecone==3.0.0
uv pip install